[INTERNAL] Add unit tests
