+++ lib/tasks/buildThemes.js
@@ -30,7 +30,7 @@ module.exports = function({workspace, dependencies, options}) {
 		promises.push(combo.byGlob(options.librariesPattern));
 	}
 
-	const compress = typeof options.compress !== "undefined" ? options.compress : true;
+	const compress = options.compress === undefined ? true : options.compress;
 
 	return Promise.all(promises).then(([allResources, availableLibraries]) => {
 		if (!availableLibraries || availableLibraries.length === 0) {
